Move test_prepared_config to statement.rs and add additional cases to it
